Fixed HA config updates to pass intent's slots and misc issues
When updating the HA config, we need to specify the slots to be passed
into the python script.

Fixed problem with script names having multiple _ in a row.  If the
generated script file name contained -, they are changed to _.  This
allowed me to notice that HA won't let you have a python script whose
name as two or more underscores in a row.

This involved changing the code to read /usr/share/snips/assistant.json
and /usr/share/snips/snippets/<dir> instead of only reading
/var/lib/snips/skills/<dir>.  Intent names might have different
capitalization from the rendered script name; the actual intent name
is found in the snippets directory; each intent is represented by a file
with an exact match name with an extension of ".snippet".  We then read
assistant.json to find the slots for the intent.  Because we're
processing the files in /usr/share/snips/snippets, we can generate the
name of the rendered script the same way "snips-template render" does.
We still copy the rendered script from /var/lib/snips/skills as before,
but this time removing any multiple _ from the final file name.