Added code to edit config.ini files for skills
Updated Web UI to have three top-level tabs, one for server control
stuff, one for logs, and one for editing config files.  Also added a
small section above the tabs to show the version of the addon and the
version of snips.

Modified extract_assistant() in funcs.sh to copy any existing config.ini
back to /share/snips so the user can edit config files even if the user
didn't initially create them.  This is the most likely case for a new
user, or when a user adds something new to their assistant for the first
time.

Fixed restart_snips_skill_server() in funcs.sh -- we need to specify
that it was started as part of snips-group.  Otherwise, it fails to
restart it.

Added some logging to ingress/control.py, since the new cheroot code
doesn't log anything.  For unknown reasons, we have to actually open the
log file from inside instead of just writing to stdout and stderr -- I
suspect cheroot is doing something with them, and I just need to fix
that.  Add support to this code for the WebUI to retrieve and save
config files.  If the config file is actually different, and is actually
used by the assistant, snips-skill-server will be restarted after it is
saved.

Updated run.sh to get version info into ingress/control.py so it can be
passed into the Web UI page.  Also disabled persistence in mosquitto,
which makes it much faster to start up.

Updated apparmor.txt for new script, and alphabetized
extract_assistant.sh script.

Updated CHANGELOG.md