Improved handling of missing assistant.zip
Modified run.sh to check if the user configured their snips console
login info.  If so, it will tell the user to use the Web UI to install
their assistant, and keep running (so the user can do that).  Otherwise,
it will exit.  The new message is a warning.

Modified messages about the missing assistant in extract_assistant
(funcs.sh) to be a warning, not error.