Updated snips_console settings and added some defaults
Changed snips console settings to be in a dictionary named
snips_console.  This separates the settings out a little better.
Fixed type of email address, since Hass.io has an "email" type.
Added panel_admin: true to add-on configuration, so the Web UI will be
limited to administrative users.

Added defaults for language and country.

Added error condition if assistant is an empty string.  The add-on
should not start in that case.  Error is logged.

Also updated documentation to match.