Added code to handle assistant training
If the assistant needs to be trained, it cannot be downloaded.  So code
has been added to handle training the assistant, the same way sam does
it.  It is then downloaded and installed.

Added code to note when the assistant is being trained and downloaded,
since both tasks take a while.  Need to add code to indicate the
assistant is being installed, too.

Also updated code to get the full path of the assistant zip the same way
we get it in the shell scripts.